requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
Flask-WTF==1.2.1
Werkzeug==3.0.1 